Make kafka dialer a public util
