+++ payment/service.go
@@ -2,10 +2,6 @@ package payment
 
 import (
 	"context"
-	"crypto/tls"
-	"crypto/x509"
-	"encoding/json"
-	"encoding/pem"
 	"io/ioutil"
 	"os"
 	"strings"
@@ -23,6 +19,7 @@ import (
 
 	"github.com/brave-intl/bat-go/utils/clients/cbr"
 	errorutils "github.com/brave-intl/bat-go/utils/errors"
+	kafkautils "github.com/brave-intl/bat-go/utils/kafka"
 	uuid "github.com/satori/go.uuid"
 	kafka "github.com/segmentio/kafka-go"
 	"github.com/shopspring/decimal"
@@ -101,96 +98,6 @@ func readFileFromEnvLoc(env string, required bool) ([]byte, error) {
 	return buf, nil
 }
 
-func tlsDialer() (*kafka.Dialer, error) {
-	keyPasswordEnv := "KAFKA_SSL_KEY_PASSWORD"
-	keyPassword := os.Getenv(keyPasswordEnv)
-
-	caPEM, err := readFileFromEnvLoc("KAFKA_SSL_CA_LOCATION", false)
-	if err != nil {
-		return nil, err
-	}
-
-	certEnv := "KAFKA_SSL_CERTIFICATE"
-	certPEM := []byte(os.Getenv(certEnv))
-	if len(certPEM) == 0 {
-		certPEM, err = readFileFromEnvLoc("KAFKA_SSL_CERTIFICATE_LOCATION", true)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	keyEnv := "KAFKA_SSL_KEY"
-	encryptedKeyPEM := []byte(os.Getenv(keyEnv))
-
-	// Check to see if KAFKA_SSL_CERTIFICATE includes both certificate and key
-	if certPEM[0] == '{' {
-		type Certificate struct {
-			Certificate string `json:"certificate"`
-			Key         string `json:"key"`
-		}
-		var cert Certificate
-		err := json.Unmarshal(certPEM, &cert)
-		if err != nil {
-			return nil, err
-		}
-		certPEM = []byte(cert.Certificate)
-		encryptedKeyPEM = []byte(cert.Key)
-	}
-
-	if len(encryptedKeyPEM) == 0 {
-		encryptedKeyPEM, err = readFileFromEnvLoc("KAFKA_SSL_KEY_LOCATION", true)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	block, rest := pem.Decode(encryptedKeyPEM)
-	if len(rest) > 0 {
-		return nil, errors.New("Extra data in KAFKA_SSL_KEY")
-	}
-
-	keyPEM := pem.EncodeToMemory(block)
-	if len(keyPassword) != 0 {
-		keyDER, err := x509.DecryptPEMBlock(block, []byte(keyPassword))
-		if err != nil {
-			return nil, errorutils.Wrap(err, "decrypt KAFKA_SSL_KEY failed")
-		}
-
-		keyPEM = pem.EncodeToMemory(&pem.Block{Type: "RSA PRIVATE KEY", Bytes: keyDER})
-	}
-
-	certificate, err := tls.X509KeyPair([]byte(certPEM), keyPEM)
-	if err != nil {
-		return nil, errorutils.Wrap(err, "Could not parse x509 keypair")
-	}
-
-	// Define TLS configuration
-	config := &tls.Config{
-		Certificates: []tls.Certificate{certificate},
-	}
-
-	// Instrument kafka cert expiration information
-	x509Cert, err := x509.ParseCertificate(certificate.Certificate[0])
-	if err != nil {
-		return nil, errorutils.Wrap(err, "Could not parse certificate")
-	}
-	kafkaCertNotBefore.Set(float64(x509Cert.NotBefore.Unix()))
-	kafkaCertNotAfter.Set(float64(x509Cert.NotAfter.Unix()))
-
-	if len(caPEM) > 0 {
-		caCertPool := x509.NewCertPool()
-		if ok := caCertPool.AppendCertsFromPEM([]byte(caPEM)); !ok {
-			return nil, errors.New("Could not add custom CA from KAFKA_SSL_CA_LOCATION")
-		}
-		config.RootCAs = caCertPool
-	}
-
-	return &kafka.Dialer{
-		Timeout:   10 * time.Second,
-		DualStack: true,
-		TLS:       config}, nil
-}
-
 // InitCodecs used for Avro encoding / decoding
 func (s *Service) InitCodecs() error {
 	s.codecs = make(map[string]*goavro.Codec)
@@ -208,11 +115,14 @@ func (s *Service) InitKafka() error {
 
 	_, logger := logging.SetupLogger(context.Background())
 
-	dialer, err := tlsDialer()
+	dialer, x509Cert, err := kafkautils.TlsDialer()
 	if err != nil {
 		return err
 	}
 	s.kafkaDialer = dialer
+
+	kafkaCertNotBefore.Set(float64(x509Cert.NotBefore.Unix()))
+	kafkaCertNotAfter.Set(float64(x509Cert.NotAfter.Unix()))
 
 	kafkaBrokers := os.Getenv("KAFKA_BROKERS")
 	kafkaWriter := kafka.NewWriter(kafka.WriterConfig{
